Written first outline for model uncertainty parameters

+++ minerva/src/base/Sample.cpp
@@ -11,7 +11,7 @@ Sample::Sample(const std::string& name, const int nbins, const double x_low, con
 	m_error.clear();
 
 	int n_anabins = GetNbinsX() + 2;
-	m_anaHist = new MnvH1D(Form("%s_nbins%.3d_anaHist", GetName(), n_anabins), "", (double)n_anabins, 0., (double)n_anabins);
+	m_anaHist = new MnvH1D(Form("%s_nbins%.3d_anaHist", GetName(), n_anabins), "", n_anabins, 0., (double)n_anabins);
 	// Do we want to inlcude under and overflow bins in this class?
 	// This makes a histogram for each interaction/reweightable var. type if we are producing splines
 }
@@ -22,7 +22,7 @@ Sample::Sample(const std::string& name, const int nbins, const double * x_bins) 
 {
 	m_error.clear();
 	int n_anabins = GetNbinsX() + 2;
-	m_anaHist = new MnvH1D(Form("%s_nbins%.3d_anaHist", GetName(), n_anabins), "", (double)n_anabins, 0, (double)n_anabins);
+	m_anaHist = new MnvH1D(Form("%s_nbins%.3d_anaHist", GetName(), n_anabins), "", n_anabins, 0, (double)n_anabins);
 }
 
 Sample::~Sample()
